Fixed Engagement index Failing to ignore obsolete "mostRecentContent"
Background a "care giver" can tell EI that there is no engagements before
the given mostRecentContent(date)  (for that "person" at that "care
giver") eg any appointments that may been in their system before that
given date are now cancelled.

According to spec any messages with a mostRecentContent before that of a
previous message should be ignored (EI already know that because the later
date already received implicitly include those engagements)

In practice there where no checks on this condition, but this has now
been implemented.

MessageCollectionStrategyImpl moved code inside a method to new method
no functional change

ProcessBean now contains this functionality/check

IncomingEngagementProcessData The process of receiving the messages deals
with both message data and the persisted entity's based on the message-
data, sorted on different attributes. This resulted in a lot of both
global and local temporary variables, this class was introduced to hold
all the process data (attempting to make the code somewhat more readable)